Distance sensor works and stops correctly PID controller is interrupt driven and distance sensor is using ping function

+++ MotorDriver/RCSketch/HCSR04.h
@@ -10,7 +10,7 @@ class HCSR04{
     int state;
 	public:
 		void setTrigger();
-		float ping();
+		long ping();
     void echoChange();
     float getInches();
     float getStart();
